examples/requirements.txt
altair
gradio
openai-agents
python-dotenv
requests
pandas
smolagents[mcp,litellm,openai,toolkit]
